Move graph walkers to separate module
