+++ tests/unit/source_code/test_graph.py
@@ -9,8 +9,8 @@ from databricks.labs.ucx.source_code.graph import (
     Dependency,
     DependencyGraph,
     InheritedContext,
-    DependencyGraphWalker,
 )
+from databricks.labs.ucx.source_code.graph_walkers import DependencyGraphWalker
 from databricks.labs.ucx.source_code.notebooks.loaders import NotebookLoader
 from databricks.labs.ucx.source_code.path_lookup import PathLookup
 from databricks.labs.ucx.source_code.python.python_ast import Tree
